nlp-resources/HTMLgetter/requirements.txt
requests
beautifulsoup4
argparse
lxml
pyyaml
selenium